+++ Sources/Base/OAuth2Base.swift
@@ -22,6 +22,9 @@ import Foundation
 #if IMPORT_SWIFT_KEYCHAIN		// experimental for SwiftKeychain integration via CocoaPods (iOS only)
 import SwiftKeychain
 #endif
+#if !NO_KEYCHAIN_IMPORT			// needs to be imported when using `swift build`, not when building via Xcode
+import SwiftKeychain
+#endif
 
 
 /// Typealias to ease working with JSON dictionaries.