+++ Sources/Base/OAuth2Error.swift
@@ -49,8 +49,10 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 	case InvalidState
 	case JSONParserError
 	case UTF8EncodeError
+	case UTF8DecodeError
 	
 	case NotUsingTLS
+	case UnableToOpenAuthorizeURL
 	case InvalidRequest
 	case UnauthorizedClient
 	case AccessDenied
@@ -124,9 +126,13 @@ public enum OAuth2Error: ErrorType, CustomStringConvertible, Equatable {
 			return "Error parsing JSON"
 		case UTF8EncodeError:
 			return "Failed to UTF-8 encode the given string"
+		case UTF8DecodeError:
+			return "Failed to decode given data as a UTF-8 string"
 		
 		case .NotUsingTLS:
 			return "You MUST use HTTPS/SSL/TLS"
+		case .UnableToOpenAuthorizeURL:
+			return "Cannot open authorize URL"
 		case .InvalidRequest:
 			return "The request is missing a required parameter, includes an invalid parameter value, includes a parameter more than once, or is otherwise malformed."
 		case .UnauthorizedClient:
@@ -169,8 +175,11 @@ public func ==(lhs: OAuth2Error, rhs: OAuth2Error) -> Bool {
 	case (.PrerequisiteFailed(let lhm), .PrerequisiteFailed(let rhm)):       return lhm == rhm
 	case (.InvalidState, .InvalidState):                         return true
 	case (.JSONParserError, .JSONParserError):                   return true
-		
-	case (.ResponseError(let lhm), .ResponseError(let rhm)):     return lhm == rhm
+	case (.UTF8EncodeError, .UTF8EncodeError):                   return true
+	case (.UTF8DecodeError, .UTF8DecodeError):                   return true
+	
+	case (.NotUsingTLS, .NotUsingTLS):                           return true
+	case (.UnableToOpenAuthorizeURL, .UnableToOpenAuthorizeURL): return true
 	case (.InvalidRequest, .InvalidRequest):                     return true
 	case (.UnauthorizedClient, .UnauthorizedClient):             return true
 	case (.AccessDenied, .AccessDenied):                         return true
@@ -178,6 +187,7 @@ public func ==(lhs: OAuth2Error, rhs: OAuth2Error) -> Bool {
 	case (.InvalidScope, .InvalidScope):                         return true
 	case (.ServerError, .ServerError):                           return true
 	case (.TemporarilyUnavailable, .TemporarilyUnavailable):     return true
+	case (.ResponseError(let lhm), .ResponseError(let rhm)):     return lhm == rhm
 	default:                                                     return false
 	}
 }