+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Version numbering represents the Swift version, plus a running number representi
 You can also refer to commit logs to get details on what was implemented, fixed and improved.
 
 
+### x.x
+
+- Move `autoDismiss` param from `authorize()` into the `authConfig` struct
+- Change `openAuthorizeURLInBrowser()` to throw instead of returning a Bool (throwing `UnableToOpenAuthorizeURL` instead of returning false)
+- Add `UTF8DecodeError`
+
+
 ### 2.1
 
 - Refresh tokens now work for all grants